+++ src/main/java/internet/shop/controllers/order/CreateOrderController.java
@@ -1,6 +1,7 @@
 package internet.shop.controllers.order;
 
 import internet.shop.lib.Injector;
+import internet.shop.model.Order;
 import internet.shop.service.OrderService;
 import internet.shop.service.ShoppingCartService;
 import java.io.IOException;
@@ -20,15 +21,14 @@ public class CreateOrderController extends HttpServlet {
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
         Long id = Long.valueOf(req.getParameter("id"));
-        Long orderId = orderService
+        Order order = orderService
                 .completeOrder(shoppingCartService
                                 .getAllProducts(shoppingCartService.getByUserId(id)),
-                        shoppingCartService.getByUserId(id).getUser())
-                .getOrderId();
+                        shoppingCartService.getByUserId(id).getUser());
         shoppingCartService.clear(shoppingCartService.get(id));
         System.out.println("List size= " + orderService.getAll().size());
-        //req.setAttribute("id",orderId);
-        //req.getRequestDispatcher("/WEB-INF/views/orders/order.jsp").forward(req,resp);
-        resp.sendRedirect(req.getContextPath() + "/orders/all");
+        req.setAttribute("order",order);
+        req.getRequestDispatcher("/WEB-INF/views/orders/order.jsp").forward(req,resp);
+        //resp.sendRedirect(req.getContextPath() + "/orders/all");
     }
 }
