+++ src/main/java/internet/shop/controllers/order/OrderController.java
@@ -15,6 +15,7 @@ public class OrderController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
+        System.out.println("asdasdsadsad");
         req.setAttribute("orders", orderService.getAll());
         req.getRequestDispatcher("/WEB-INF/views/orders/all.jsp").forward(req, resp);
     }
