+++ src/main/java/internet/shop/controllers/shoppingcart/CreateShopCartController.java
@@ -1,28 +1,27 @@
-package internet.shop.controllers.shoppingсart;
+package internet.shop.controllers.shoppingcart;
 
 import internet.shop.lib.Injector;
-import internet.shop.model.Product;
+import internet.shop.model.ShoppingCart;
 import internet.shop.service.ShoppingCartService;
+import internet.shop.service.UserService;
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class ShoppingCartController extends HttpServlet {
+public class CreateShopCartController extends HttpServlet {
     private static final Long USER_ID = 0L;
     private static final Injector injector = Injector.getInstance("internet.shop");
     private ShoppingCartService shoppingCartService = (ShoppingCartService) injector
             .getInstance(ShoppingCartService.class);
+    private UserService userService = (UserService) injector.getInstance(UserService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Product> items = shoppingCartService.getByUserId(USER_ID).getItems();
-        req.setAttribute("shoppingCartItems", items);
-        req.setAttribute("userId",USER_ID);
-        req.getRequestDispatcher("/WEB-INF/views/shoppingCart/shoppingCartItems.jsp")
-                .forward(req,resp);
+        ShoppingCart shoppingCart = new ShoppingCart(userService.get(USER_ID));
+        shoppingCartService.create(shoppingCart);
+        resp.sendRedirect("http://localhost:8080/shoppingCart");
     }
 }
