+++ src/main/java/internet/shop/controllers/shoppingcart/ShoppingCartController.java
@@ -1,28 +1,28 @@
-package internet.shop.controllers.shoppingсart;
+package internet.shop.controllers.shoppingcart;
 
 import internet.shop.lib.Injector;
-import internet.shop.service.ProductService;
+import internet.shop.model.Product;
 import internet.shop.service.ShoppingCartService;
 import java.io.IOException;
+import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class AddProductToShopCartController extends HttpServlet {
+public class ShoppingCartController extends HttpServlet {
     private static final Long USER_ID = 0L;
     private static final Injector injector = Injector.getInstance("internet.shop");
     private ShoppingCartService shoppingCartService = (ShoppingCartService) injector
             .getInstance(ShoppingCartService.class);
-    private ProductService productService = (ProductService) injector
-            .getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Long productId = Long.valueOf(req.getParameter("productId"));
-        shoppingCartService.addProduct(shoppingCartService.getByUserId(USER_ID),
-                productService.get(productId));
-        resp.sendRedirect(req.getContextPath() + "/products/all;");
+        List<Product> items = shoppingCartService.getByUserId(USER_ID).getItems();
+        req.setAttribute("shoppingCartItems", items);
+        req.setAttribute("userId",USER_ID);
+        req.getRequestDispatcher("/WEB-INF/views/shoppingCart/shoppingCartItems.jsp")
+                .forward(req,resp);
     }
 }
