+++ src/main/webapp/web.xml
@@ -69,7 +69,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>shoppingCart</servlet-name>
-        <servlet-class>internet.shop.controllers.shoppingсart.ShoppingCartController</servlet-class>
+        <servlet-class>internet.shop.controllers.shoppingcart.ShoppingCartController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>shoppingCart</servlet-name>
@@ -77,7 +77,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>addProductToShopCart</servlet-name>
-        <servlet-class>internet.shop.controllers.shoppingсart.AddProductToShopCartController</servlet-class>
+        <servlet-class>internet.shop.controllers.shoppingcart.AddProductToShopCartController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>addProductToShopCart</servlet-name>
@@ -85,7 +85,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>removeProductFromShopCart</servlet-name>
-        <servlet-class>internet.shop.controllers.shoppingсart.RemoveProductFromShopCartController</servlet-class>
+        <servlet-class>internet.shop.controllers.shoppingcart.RemoveProductFromShopCartController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>removeProductFromShopCart</servlet-name>
@@ -93,7 +93,7 @@
     </servlet-mapping>
     <servlet>
         <servlet-name>CreateShopCart</servlet-name>
-        <servlet-class>internet.shop.controllers.shoppingсart.CreateShopCartController</servlet-class>
+        <servlet-class>internet.shop.controllers.shoppingcart.CreateShopCartController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>CreateShopCart</servlet-name>
